moved bobadex to realms, made realm id required in bobadex query
