+++ server/realms/sql/fetch-bobadex.sql
